backend/requirements/common.txt
django==3.2
django-baton
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
django-cors-headers
Pillow
requests
django-dotenv